pyproject.toml
[project]
name = "library-online"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "black>=25.1.0",
    "fastapi-cache2[redis]>=0.2.2",
    "fastapi[standard]>=0.115.12",
    "mypy>=1.15.0",
    "passlib>=1.7.4",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "pyright>=1.1.399",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "pytest-dotenv>=0.5.2",
    "ruff>=0.11.6",
    "sqlalchemy>=2.0.40",
]

[tool.ruff]
line-length = 100
